requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.40.0
numpy>=1.24.0
pandas>=2.0.0

# NLP & Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13

# Retrieval
scikit-learn>=1.3.0
faiss-cpu>=1.8.0  # Use faiss-gpu if CUDA is available

# Utilities
tqdm>=4.65.0

# Optional: GPU support (uncomment if using CUDA)
# faiss-gpu>=1.8.0
